feat: implement KadDHT analysis logic and tracer for detailed DHT lookup metrics
